// client/packages/barda-design/src/components/button.tsx
export const BigButtonStyle = styled.button`
  height: 30px;
  width: 280px;
  margin-right: 2px;
  color: #4965f2;
  border: 1px solid #c9d1fc;
  background: #fafbff;
  border-radius: 4px;

  font-size: 13px;
  text-align: center;
  line-height: 13px;
  transition: all 0.4s ease;
  cursor: pointer;

  &:hover {
    color: #3377ff;
    border: 1px solid #c2d6ff;
    background: #f9fbff;
    transition: all 0.4s ease;
  }
`;